Recommit: "[llvm-exegesis] Improve error reporting in Assembler.cpp"
Summary: Commit 543624b616424ac389256472761ae4e907361e9c was reverted in
bbf42b2b95f0e5ea582492f32be2a28632526065 because it broke builds that treat
warnings as errors.
This commit updates the calls to `assembleToStream()` in tests to check that
the return value is valid.

Original commit message:

Followup to D74084.
Replace the use of `report_fatal_error()` with returning the error to
`llvm-exegesis.cpp` and handling it there.

Differential Revision: https://reviews.llvm.org/D74325