[X86] Teach combineToExtendBoolVectorInReg to create opportunities for using broadcast load instructions.
If we're inserting a scalar that is smaller than the element
size of the final VT, the value of the extra bits doesn't matter.

Previously we any_extended in the scalar domain before inserting.

This patch changes this to use a broadcast of the original
scalar type and then a bitcast to the final type. This might
enable the use of a broadcast load.

This recovers regressions from 1939da52f6e48386653a299fe8f25709a49a3c14
and 0ad738d532ebfef9f5fc30f83dff0ed18f441127 without relying on
alignment of the load.

Differential Revision: https://reviews.llvm.org/D75835