Revert "[InstCombine] do not exclude min/max from icmp with casted operand fold"
This reverts commit 5b36c57220fe273ae29ca633f07ee7e197d7d1b7.
As noted in the post-commit thread:
https://reviews.llvm.org/rGf2fbdf76d8d0
...this can obscure a min/max pattern where the components
have extra uses. We can show that the problem is independent
of this change with a slightly modified source example, so
this revert just delays/reduces the need to fix the real
problem.

We need to improve our analysis of negation or -- more
generally -- subtraction using patches like D77230 or D68408.