[PowerPC][Future] Fix for 2e09247736d4466c1f3ededa807cf045d93e67c4
This is a fix for the previous patch 2e09247736d4466c1f3ededa807cf045d93e67c4.
In some cases it may be possible to have the compiler produce st_other=1 without
the compiler using mcpu=future which should not be the case. This patch adds a
guard to make sure that if we are using st_other=1 then we are also compiling
for future CPU.